refactor getLibrary to not use maker provider
