+++ lib/maker/index.ts
@@ -3,7 +3,7 @@ import McdPlugin, { DAI } from '@makerdao/dai-plugin-mcd';
 import LedgerPlugin from '@makerdao/dai-plugin-ledger-web';
 import TrezorPlugin from '@makerdao/dai-plugin-trezor-web';
 import GovernancePlugin, { MKR } from '@makerdao/dai-plugin-governance';
-import { Web3ReactPlugin } from './web3react';
+import { Web3ReactPlugin } from '../web3react';
 
 import { SupportedNetworks, DEFAULT_NETWORK } from '../constants';
 import { networkToRpc } from './network';
