+++ lib/web3react/connectors.ts
@@ -1,22 +1,9 @@
-// TODO move everything under web3react/ to its own module
-
-import ProviderSubprovider from 'web3-provider-engine/dist/es5/subproviders/provider';
 import { InjectedConnector } from '@web3-react/injected-connector';
 import { WalletConnectConnector, WalletConnectConnectorArguments } from '@web3-react/walletconnect-connector';
 import { WalletLinkConnector } from '@web3-react/walletlink-connector';
 import { AbstractConnector } from '@web3-react/abstract-connector';
-import { networkToRpc } from '../network';
-import { SupportedNetworks } from '../../constants';
-
-export const Web3ReactPlugin = maker => {
-  maker.service('accounts', true).addAccountType('web3-react', ({ library, address }) => {
-    const { provider, connector } = library;
-    const subprovider = new ProviderSubprovider(provider);
-    return { subprovider, address, connector };
-  });
-};
-
-export const getLibrary = (provider, connector) => ({ provider, connector });
+import { networkToRpc } from '../maker/network';
+import { SupportedNetworks } from '../constants';
 
 const POLLING_INTERVAL = 12000;
