+++ lib/web3react/hooks.ts
@@ -6,7 +6,7 @@
 import { useContext, useEffect, useState } from 'react';
 import { useWeb3React } from '@web3-react/core';
 
-import getMaker from '../index';
+import getMaker from '../maker/index';
 import { injectedConnector } from './index';
 import { AnalyticsContext } from 'modules/app/client/analytics/AnalyticsContext';
