+++ package.json
@@ -45,6 +45,7 @@
     "date-fns": "^2.24.0",
     "date-fns-tz": "^1.1.6",
     "ethereum-ens": "^0.8.0",
+    "ethers": "^5.5.2",
     "gray-matter": "^4.0.2",
     "ipfs-only-hash": "^2.0.1",
     "lodash": "^4.17.21",
